Moved VariableNameUtils into org.openrewrite.java.
