+++ rewrite-test/src/main/kotlin/org/openrewrite/java/VariableNameUtilsTest.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openrewrite.java.search
+package org.openrewrite.java
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
@@ -23,9 +23,6 @@ import org.openrewrite.Cursor
 import org.openrewrite.ExecutionContext
 import org.openrewrite.InMemoryExecutionContext
 import org.openrewrite.Issue
-import org.openrewrite.java.JavaIsoVisitor
-import org.openrewrite.java.JavaParser
-import org.openrewrite.java.internal.VariableNameUtils
 import org.openrewrite.java.tree.J
 import org.openrewrite.test.RewriteTest
 import org.openrewrite.test.RewriteTest.toRecipe
